requirements.txt
pydantic
boto3
PyMuPDF
python-dotenv